+++ cds-feature-attachments/src/test/java/com/sap/cds/feature/attachments/handler/applicationservice/processor/readhelper/AttachmentStatusExceptionTest.java
@@ -3,7 +3,6 @@ package com.sap.cds.feature.attachments.handler.applicationservice.processor.rea
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import com.sap.cds.feature.attachments.handler.applicationservice.processor.readhelper.AttachmentStatusException;
 import com.sap.cds.feature.attachments.utilities.AttachmentErrorStatuses;
 
 class AttachmentStatusExceptionTest {
